+++ src/main/java/br/edu/ifpb/praticas/sparta/entidades/Pesquisa.java
@@ -15,26 +15,26 @@ public class Pesquisa implements Serializable{
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
     @Column(nullable = false, unique = true)
-    private int atendimento;
+    private long atendimento;
     @Column(columnDefinition = "BOOLEAN DEFAULT FALSE")
     private boolean respondido;
     private float nota;
 
     public Pesquisa(){ }
 
-    public Pesquisa(int id, int atendimento, boolean respondido, float nota){
+    public Pesquisa(long id, long atendimento, boolean respondido, float nota){
         this.id = id;
         this.atendimento = atendimento;
         this.respondido = respondido;
         this.nota = nota;
     }
 
-    public int getId() {return id;}
-    public void setId(int id) {this.id = id;}
-    public int getAtendimento() {return atendimento;}
-    public void setAtendimento(int atendimento) {this.atendimento = atendimento;}
+    public long getId() {return id;}
+    public void setId(long id) {this.id = id;}
+    public long getAtendimento() {return atendimento;}
+    public void setAtendimento(long atendimento) {this.atendimento = atendimento;}
     public boolean isRespondido() {return respondido;}
     public void setRespondido(boolean respondido) {this.respondido = respondido;}
     public float getNota() {return nota;}
